# openapi.json
{"openapi":"3.1.0","info":{"title":"Q Arena API","description":"FastAPI backend: subject search, published data access, answer generation, optional student uploads.","version":"2.0.0"},"paths":{"/subjects/search":{"get":{"tags":["Student Subjects"],"summary":"Search Subjects","operationId":"search_subjects_subjects_search_get","parameters":[{"name":"query","in":"query","required":true,"schema":{"type":"string","title":"Query"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubjectSearchResult"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/subjects/{subject_code}/overview":{"get":{"tags":["Student Subjects"],"summary":"Get Subject Overview","operationId":"get_subject_overview_subjects__subject_code__overview_get","parameters":[{"name":"subject_code","in":"path","required":true,"schema":{"type":"string","title":"Subject Code"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubjectOverviewResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/subjects/{subject_code}/questions":{"get":{"tags":["Student Subjects"],"summary":"List Subject Questions","operationId":"list_subject_questions_subjects__subject_code__questions_get","parameters":[{"name":"subject_code","in":"path","required":true,"schema":{"type":"string","title":"Subject Code"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response List Subject Questions Subjects  Subject Code  Questions Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/subjects/{subject_code}/analysis":{"get":{"tags":["Student Subjects"],"summary":"Get Subject Analysis","operationId":"get_subject_analysis_subjects__subject_code__analysis_get","parameters":[{"name":"subject_code","in":"path","required":true,"schema":{"type":"string","title":"Subject Code"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Get Subject Analysis Subjects  Subject Code  Analysis Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/subjects/{subject_code}/prediction":{"get":{"tags":["Student Subjects"],"summary":"Get Subject Prediction","operationId":"get_subject_prediction_subjects__subject_code__prediction_get","parameters":[{"name":"subject_code","in":"path","required":true,"schema":{"type":"string","title":"Subject Code"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Get Subject Prediction Subjects  Subject Code  Prediction Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/search":{"post":{"tags":["Student Search"],"summary":"Semantic Search","operationId":"semantic_search_search_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SearchRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Semantic Search Search Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/answers/generate":{"post":{"tags":["Student Answers"],"summary":"Generate Exam Answer","operationId":"generate_exam_answer_answers_generate_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AnswerGenerateRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AnswerGenerateResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/student/upload-question-paper":{"post":{"tags":["Student API - Upload"],"summary":"Upload Student Question Paper","description":"Allow students to upload their own question paper.\nThese uploads are marked as draft/student source and not used in published analysis.","operationId":"upload_student_question_paper_student_upload_question_paper_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StudentExamUploadRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"anyOf":[{"type":"integer"},{"type":"string"}]},"type":"object","title":"Response Upload Student Question Paper Student Upload Question Paper Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/admin/exams":{"post":{"tags":["Admin Exams"],"summary":"Create Exam Admin","operationId":"create_exam_admin_admin_exams_post","requestBody":{"content":{"application/json":{"schema":{"anyOf":[{"additionalProperties":true,"type":"object"},{"items":{"additionalProperties":true,"type":"object"},"type":"array"}],"title":"Payload"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Create Exam Admin Admin Exams Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/admin/subjects/{subject_code}/publish":{"post":{"tags":["Admin Exams"],"summary":"Publish Subject Endpoint","operationId":"publish_subject_endpoint_admin_subjects__subject_code__publish_post","parameters":[{"name":"subject_code","in":"path","required":true,"schema":{"type":"string","title":"Subject Code"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":{"anyOf":[{"type":"integer"},{"type":"string"}]},"title":"Response Publish Subject Endpoint Admin Subjects  Subject Code  Publish Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/admin/extract/upload":{"post":{"tags":["Admin Extraction"],"summary":"Upload Pdf","operationId":"upload_pdf_admin_extract_upload_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_pdf_admin_extract_upload_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Upload Pdf Admin Extract Upload Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/admin/extract/result/{file_id}":{"get":{"tags":["Admin Extraction"],"summary":"Get Extraction Result","operationId":"get_extraction_result_admin_extract_result__file_id__get","parameters":[{"name":"file_id","in":"path","required":true,"schema":{"type":"string","title":"File Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Get Extraction Result Admin Extract Result  File Id  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/admin/extract/save-reviewed":{"post":{"tags":["Admin Extraction"],"summary":"Save Reviewed Result","operationId":"save_reviewed_result_admin_extract_save_reviewed_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReviewedExtractionPayload"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Save Reviewed Result Admin Extract Save Reviewed Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"tags":["Health"],"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object","title":"Response Health Check Health Get"}}}}}}}},"components":{"schemas":{"AnswerGenerateRequest":{"properties":{"question":{"type":"string","title":"Question"},"subject_code":{"type":"string","title":"Subject Code"},"answer_type":{"type":"string","title":"Answer Type","default":"5_mark"}},"type":"object","required":["question","subject_code"],"title":"AnswerGenerateRequest"},"AnswerGenerateResponse":{"properties":{"question":{"type":"string","title":"Question"},"subject_code":{"type":"string","title":"Subject Code"},"answer_type":{"type":"string","title":"Answer Type"},"answer":{"type":"string","title":"Answer"},"context_questions":{"items":{"type":"string"},"type":"array","title":"Context Questions"}},"type":"object","required":["question","subject_code","answer_type","answer","context_questions"],"title":"AnswerGenerateResponse"},"Body_upload_pdf_admin_extract_upload_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_pdf_admin_extract_upload_post"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"QuestionCreate":{"properties":{"question_no":{"type":"string","title":"Question No"},"question_text":{"type":"string","title":"Question Text"},"marks":{"type":"integer","minimum":0.0,"title":"Marks"},"topic":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Topic"}},"type":"object","required":["question_no","question_text","marks"],"title":"QuestionCreate"},"ReviewedExtractionPayload":{"properties":{"file_id":{"type":"string","minLength":1,"title":"File Id"},"data":{"additionalProperties":true,"type":"object","title":"Data"}},"type":"object","required":["file_id","data"],"title":"ReviewedExtractionPayload"},"SearchRequest":{"properties":{"query":{"type":"string","title":"Query"},"subject_code":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Subject Code"},"top_k":{"type":"integer","maximum":20.0,"minimum":1.0,"title":"Top K","default":5}},"type":"object","required":["query"],"title":"SearchRequest"},"StudentExamUploadRequest":{"properties":{"exam_name":{"type":"string","title":"Exam Name"},"exam_year":{"type":"integer","title":"Exam Year"},"subject_code":{"type":"string","title":"Subject Code"},"time":{"type":"string","title":"Time"},"total_marks":{"type":"integer","minimum":0.0,"title":"Total Marks"},"questions":{"items":{"$ref":"#/components/schemas/QuestionCreate"},"type":"array","minItems":1,"title":"Questions"}},"type":"object","required":["exam_name","exam_year","subject_code","time","total_marks","questions"],"title":"StudentExamUploadRequest"},"SubjectOverviewResponse":{"properties":{"found":{"type":"boolean","title":"Found"},"subject_code":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Subject Code"},"subject_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Subject Name"},"total_questions":{"type":"integer","title":"Total Questions","default":0},"available_years":{"items":{"type":"integer"},"type":"array","title":"Available Years","default":[]},"top_topics":{"items":{"$ref":"#/components/schemas/TopicSummary"},"type":"array","title":"Top Topics","default":[]},"prediction_available":{"type":"boolean","title":"Prediction Available","default":false},"message":{"type":"string","title":"Message","default":""}},"type":"object","required":["found"],"title":"SubjectOverviewResponse"},"SubjectSearchResult":{"properties":{"found":{"type":"boolean","title":"Found"},"subject_code":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Subject Code"},"subject_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Subject Name"},"total_questions":{"type":"integer","title":"Total Questions","default":0},"available_years":{"items":{"type":"integer"},"type":"array","title":"Available Years","default":[]},"can_upload":{"type":"boolean","title":"Can Upload","default":true},"message":{"type":"string","title":"Message","default":""}},"type":"object","required":["found"],"title":"SubjectSearchResult"},"TopicSummary":{"properties":{"topic":{"type":"string","title":"Topic"},"frequency":{"type":"integer","title":"Frequency"},"sample_questions":{"items":{"type":"string"},"type":"array","title":"Sample Questions"}},"type":"object","required":["topic","frequency","sample_questions"],"title":"TopicSummary"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}